requirements.txt
multipledispatch
pyyaml
pytest
pytest-asyncio
pydantic
markitdown
openai
aiohttp
playwright
pre-commit
duckdb
pandas
networkx
pyarrow
psutil
azure-ai-documentintelligence